Per review, move scatterPlots() out of scatter-plot.js, refine name
