+++ app/javascript/lib/scatter-plot.js
@@ -14,7 +14,6 @@ import Plotly from 'plotly.js-dist'
 
 import { labelFont, getColorBrewerColor } from 'lib/plot'
 import { fetchCluster } from 'lib/scp-api'
-import { getMainViewOptions } from 'lib/study-overview/view-options'
 
 // List of raw plots (API data + UI props, but not yet Plotly-processed)
 let scatterPlots = []
@@ -285,45 +284,4 @@ export async function scatterPlot(apiParams, props) {
   renderScatterPlot(rawPlot, plotId, legendId)
 
   spinner.stop()
-}
-
-/**
- * Load and draw scatter plots for reference clusters, gene + ref clusters
- *
- * @param {Object} study Study object returned by clusters API endpoint
- * @param {String} gene Searched gene name, e.g. "TP53"
- * @param {Boolean} hasReference Whether each plot has a paired reference
- */
-export async function initScatterPlots(study, gene=null, hasReference=false) {
-  const baseSelector = '#scatter-plots .panel-body'
-  const $container = $(baseSelector)
-  $container.html('<div class="row multiplot"></div>')
-  if (hasReference) {$container.append('<div class="row multiplot"></div>')}
-
-  const accession = study.accession
-
-  // Plot UI properties that are distinct from fetched scatter plot data
-  const props = {
-    numRows: (hasReference ? 2 : 1),
-    numColumns: (study.spatialGroupNames.length > 0) ? 2 : 1
-  }
-
-  for (let i = 0; i < props.numColumns; i++) {
-    const options = getMainViewOptions(i)
-    const apiParams = Object.assign({ accession, gene }, options)
-
-    props.selector = baseSelector + ' .multiplot:nth-child(1)'
-    props.hasLegend = true
-    props.plotId = `scatter-plot-${i}`
-
-    scatterPlot(apiParams, props)
-
-    if (hasReference) {
-      apiParams.gene = null
-      props.selector = baseSelector + ' .multiplot:nth-child(2)'
-      props.hasLegend = false
-      props.plotId += '-reference'
-      scatterPlot(apiParams, props)
-    }
-  }
 }
